Downloads/share_smote/share_smote/share_smote/backend_python/requirements.txt
fastapi
uvicorn[standard]
pillow
scikit-image
scikit-learn
imbalanced-learn
python-multipart
pydantic
pydantic-settings
pathspec
filelock
python-jose[cryptography]
passlib[bcrypt]
slowapi
pytest
requests